Adding Ingredient tests
